fix(transforms): close 6 review gaps before merging native-transform-ownership
- A push always wins over a binding: set_world_transform now unbinds the
  store slot by construction, so a future explicit-matrix push can no
  longer be silently overwritten by the next frame's store sweep. Fixed
  a regression this exposed in set_instance_transform_slot's own
  "compose once now" step (it was unbinding what it had just bound).
- Deleted the false/harmful TransformStore reset in conftest.py — the
  native store is a process-wide singleton that is never reset; the
  Python backend actively broke on it.
- Added a test that forces get_store()'s Python-backend branch and
  round-trips a real ObjectClass through it end-to-end.
- Widened NativeTransformStore's exception catch to also convert a
  pybind11 TypeError (negative index) into StaleHandleError.
- Renamed the _debug_sync_instance_transforms test hook to
  _test_only_sync_instance_transforms (gating behind developer mode
  isn't reachable from pytest, so it would have silently no-op'd).
- Carried transform-store scale as double end-to-end through
  compose_world_matrix so it rounds to float once, after the rotation
  multiply, matching _world_matrix_from's double-precision arithmetic.

+++ native/src/scenegraph/include/scenegraph/instance.h
@@ -42,7 +42,11 @@ struct Instance {
     /// every render-interpolated ship still take that path).
     int           xform_index = -1;
     std::uint32_t xform_generation = 0;
-    float         xform_scale = 1.0f;
+    // double, not float: compose_world_matrix multiplies this by the store's
+    // double-precision rotation and rounds to float only once, at the end —
+    // matching engine/host_loop.py:_world_matrix_from. A float here would
+    // round the scale before the multiply instead of after.
+    double        xform_scale = 1.0;
 
     bool visible = true;
     Pass pass = Pass::Space;
